rag/level-1/requirements.txt
# Core dependencies for PDF RAG Search System
# ============================================

# Azure OpenAI integration
openai>=1.0.0

# PDF processing
PyPDF2>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# Scientific computing and ML
numpy>=1.24.0
scikit-learn>=1.3.0

# PDF report generation
reportlab>=4.0.0

# Type hints support
typing-extensions>=4.0.0

# File operations and system utilities
glob2>=0.7

# JSON processing (built-in, but ensuring compatibility)
# json - built-in module

# Date and time handling (built-in, but ensuring compatibility)
# datetime - built-in module

# System operations (built-in, but ensuring compatibility)
# os - built-in module
# subprocess - built-in module
# platform - built-in module
# time - built-in module

# Optional: Development and testing dependencies
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
